requirements.txt
Pillow==11.0.0
streamlit==1.40.0
